Finished implementation of job dependecy list on slurmd and slurmstepd, new function to exclude jobs too much penalized out of mates selection

+++ src/slurmctld/slurmctld.h
@@ -771,7 +771,7 @@ struct job_record {
 	time_t wait4switch_start; /* Time started waiting for switch       */
 
 	/* Marco: my parameters*/
-	int penalty;                    /* Penalization due to malleability */
+	double penalty;                    /* Penalization due to malleability */
 };
 
 /* Job dependency specification, used in "depend_list" within job_record */
